requirements.txt
streamlit>=1.32.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0
faiss-cpu>=1.8.0
pypdf>=4.0.0
python-dotenv>=1.0.0
openai>=1.20.0
tiktoken>=0.7.0
sentence-transformers>=3.0.0
torch>=2.0.0
sqlalchemy>=2.0.0
docx2txt>=0.8
torchvision